dix delete bug, exchange places of passed params

+++ pages/api/wishlist.js
@@ -6,14 +6,14 @@ export default async function handler(req, res) {
   // Add to Wishlist
   if (req.method === 'POST') {
     // Extract data from the request body
-    const { customerId, productId } = req.body;
-    console.log("userId", customerId)
+    const { userId, productId } = req.body;
+    console.log("userId", userId, productId)
     try {
       // Create a new wishlist entry using Prisma
       await prisma.wishlist.create({
         data: {
-          customerId: parseInt(customerId), // Set the customerId directly
-          productId: parseInt(productId), // Set the productId directly
+          customerId: Number(productId), 
+          productId: Number(userId), 
         },
       });
 
@@ -31,18 +31,45 @@ export default async function handler(req, res) {
 
   if (req.method === 'DELETE') {
     try {
-      const { customerId, productId } = req.body;
-
+      const { userId, productId } = req.body;
+      console.log("userId, productId SERVER", userId, productId )
       // Call the deleteFromWishlist function with customerId and productId
-      const result = await prisma.wishlist.delete({
+      const recordToDelete = await prisma.wishlist.findFirst({
         where: {
-          customerId: parseInt(customerId),
-          productId: parseInt(productId),
+          customerId: Number(userId),
+          productId: Number(productId),
         },
+        select: {
+          id: true
+        }
       });
 
-      // Respond with a success message or status
-      res.status(200).json(result);
+      // const recordToDelete = await prisma.wishlist.findFirst({
+      //   where: {
+      //     customerId: 6,
+      //   }
+      // });
+      console.log("recordToDelete", recordToDelete )
+      if (recordToDelete) {
+        await prisma.wishlist.delete({
+          where: {
+            id: recordToDelete.id,
+          },
+        });
+        const result = await prisma.wishlist.findMany({
+          where: {
+            customerId: Number(userId),
+          },
+          include: {
+            product: true,
+            // Add any other fields you want to include from the related models.
+          },
+        });
+        // Respond with a success message or status
+        res.status(200).json(result);
+      }
+
+      
     } catch (error) {
       // Handle errors, e.g., item not found
       console.error('Error deleting item from wishlist:', error);
